Eliminated un needed floating point casts

+++ src/fileaccess.cpp
@@ -798,7 +798,7 @@ bool FileAccess::interruptableReadFile(void* pDestBuffer, qint64 maxLength)
         }
         i += reallyRead;
 
-        ProgressProxy::setCurrent(floor(double(i) / maxLength * 100));
+        ProgressProxy::setCurrent(std::floor(i / maxLength * 100));
         if(ProgressProxy::wasCancelled())
             return false;
     }
